Improved shell command interpreter

+++ Userland/SampleCodeModule/include/phylo.h
@@ -1,9 +1,6 @@
 #ifndef _PHYLO_H_
 #define _PHYLO_H_
 
-/* Standard library */
-#include <stdint.h>
-
-void phyloProcess(uint64_t argc, char* argv[]);
+void phyloProcess(int argc, char* argv[]);
 
 #endif
