+++ Userland/SampleCodeModule/phylo.c
@@ -3,6 +3,6 @@
 #include <userstdlib.h>
 #include <phylo.h>
 
-void phyloProcess(uint64_t argc, char* argv[]) {
+void phyloProcess(int argc, char* argv[]) {
 
 }
